requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.1
certifi==2025.8.3
charset-normalizer==3.4.2
choreographer==1.2.1
click==8.2.1
colorama==0.4.6
defusedxml==0.7.1
fastapi==0.118.0
feedparser==6.0.12
Flask==3.1.1
flet==0.28.3
fonttools==4.61.0
fpdf2==2.8.5
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.45
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaleido==1.2.0
logistro==2.0.1
MarkupSafe==3.0.2
multidict==6.7.1
mysql-connector==2.2.9
mysql-connector-python==9.5.0
narwhals==2.10.1
numpy==2.3.4
oauthlib==3.3.1
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.0
pyarrow==21.0.0
pydantic==2.11.10
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
repath==0.9.0
requests==2.32.4
rpds-py==0.28.0
schedule==1.2.2
sgmllib3k==1.0.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.48.0
streamlit==1.51.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
twilio==9.10.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
Werkzeug==3.1.3
yarl==1.22.0
